// src/api/models/OAuth2TokenRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OAuth2TokenRequest = {
	grantType?: string | null;
	clientId?: string | null;
	clientSecret?: string | null;
	username?: string | null;
	password?: string | null;
	refreshToken?: string | null;
	scope?: string | null;
	rememberMe?: boolean | null;
	signature?: string | null;
	signatureAlgorithm?: string | null;
};